requirements.txt
pydantic
contextvars>=2.4
uvicorn>=0.46.0
fastapi>=0.110.0
granian>=1.2.0
qqmusic-api-python>=0.1.0
httpx>=0.27.0
pydantic-settings>=2.2.0